% Generated by build script
\name{psmsl_annual}
\alias{psmsl_annual}
\title{Fetch annual mean sea level data for a PSMSL station}
\usage{
psmsl_annual()
}
\arguments{
\item{station_id}{PSMSL station ID (from psmsl_stations)}
}
\value{tibble: year (integer), level_mm (numeric, mm above RLR datum),}
\description{Fetch annual mean sea level data for a PSMSL station}
